whyred: Add interface specification for lights hal.
So that init knows how to start it and to provide
an example of how to specify interfaces since there
are no lazy starting services on Pixel yet.

In order to make lights a lazy service, it would also
have to be marked as 'oneshot'.

Test: manually using Light as a lazy service
Bug: 64678982
Merged-In: I846a7e2d9545b42af9b985adb74883bda1b8c355
Change-Id: I846a7e2d9545b42af9b985adb74883bda1b8c355